backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
langchain==0.1.12
langchain-openai==0.0.8
python-dotenv==1.0.1
gunicorn==21.2.0
langchain-community==0.3.10